backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
# Optional: For advanced AI chatbot (uncomment if you want to use Llama 2)
# transformers==4.36.0
# torch==2.1.0
# supabase==2.3.0
